conversion to 32 bit ints

+++ NeuronAnalysis/fit_learning_model.py
@@ -531,11 +531,11 @@ def fit_learning_rates(NN_FIT, blocks, trial_sets, learn_t_win=None, bin_width=1
     lower_bounds = [x[1][1] for x in sorted(param_conds.items(), key=lambda item: item[1][3])]
     upper_bounds = [x[1][2] for x in sorted(param_conds.items(), key=lambda item: item[1][3])]
     W_min_pf = np.float64(0.0)
-    FR_MAX = lf_kwargs['FR_MAX']
-    tau_rise_CS = lf_kwargs['tau_rise_CS']
-    tau_decay_CS = lf_kwargs['tau_decay_CS']
-    tau_rise_CS_LTP = lf_kwargs['tau_rise_CS_LTP']
-    tau_decay_CS_LTP = lf_kwargs['tau_decay_CS_LTP']
+    FR_MAX = np.int32(lf_kwargs['FR_MAX'])
+    tau_rise_CS = np.int32(lf_kwargs['tau_rise_CS'])
+    tau_decay_CS = np.int32(lf_kwargs['tau_decay_CS'])
+    tau_rise_CS_LTP = np.int32(lf_kwargs['tau_rise_CS_LTP'])
+    tau_decay_CS_LTP = np.int32(lf_kwargs['tau_decay_CS_LTP'])
     # Finally append CS to inputs and get other args needed for learning function
     fit_inputs = np.hstack([bin_eye_data, binned_CS[:, None]])
